requirements.txt
openai
pytesseract
opencv-python
Pillow
python-dotenv
pdf2image
